Add endpoint to generate JWT tokens for MQTT authentication
This also introduces a class to hold application configuration; for now there's just the secret key used for JWT signing.

It includes the Micronaut JWT authentication library, which may be a good option for users' login sessions but requires some experimentation.

Example command to generate a token with this endpoint and pass it to the HiveMQ MQTT command-client client to connect to an instance of Mosquitto with the JWT auth plugin running on a host at 192.168.1.73:

    mqtt sub -h 192.168.1.73 -p 1883 -u 1 \
        --password "$(curl -s -X POST -H "Authorization: Basic $(echo -n user:dummyKey | base64)" http://localhost:8080/api/v1/token | jq -r .token)" \
         -t '1/#' -ws -d -v

CU-hwvmzf