Implement MQTT communication with device manager
Move the Mosquitto credentials endpoint to a path that more explicitly says what it's for; this matches the path that is now used by rhizo-client when it is configured to get MQTT credentials from the server.

Add rhizo-server-compatible endpoints to test whether a timeseries (sequence) exists, and to create a timeseries if it doesn't already exist.

Parse sequence updates and log messages as generated by the device manager and turn them into application events. This commit doesn't include any of the event handlers, just the low-level message handling.

Add some demo values to the device table to match the configuration of the "sim" site in the sites repository so we can run the sim site and the values will be added to the database.