Add "mark notification as read" endpoints
Define `/api/v1/seedbank/notification/X/markRead` where X is either a notification
ID or the magic string `all` to mark all notifications as read.

As part of that, define some standard response formats for write operations and
errors, so clients can look in a standard place to get error messages and so on.