Sort endpoints and schemas in Swagger YAML document
By default, the endpoints and schemas are in an undefined order, which means they could move around at any time. Change it to sort both in lexical order so diffs between versions of the document are more reliably useful.

This commit updates the document to use the new sort order, so the document has a bunch of edits here, but doesn't change any of the API contracts.