Generate enums for reference tables
Values such as accession state affect back end business logic, so treating them as
integer IDs or strings isn't ideal. Add code generation logic to turn a select set
of tables into Kotlin enum classes.

This does not cover all the reference tables, just the ones that the back end code
will need to use as part of its logic. Other reference tables are only used as a
way to tell the client a list of possible values to show in a drop-down menu;
those continue to be treated as regular tables.