Use database for notification endpoints
Populate the database with some test notifications and change the notification endpoints to hit the database. Add a "mark all as unread" endpoint (only available in dev) for test purposes.

Switch the generated code to use `Instant` instead of `OffsetDateTime` for timestamps; `OffsetDateTime` required some extra boilerplate code and we're not preserving offsets anyway.

CU-kmztdy