Initial CRUD endpoints for accessions
Adds endpoints to create, read, and update (but not delete) accessions. This is
still a bit of a work in progress but should be functional enough for the front
end team to start developing against.

Since accession data is spread across multiple tables in the database, there is
some actual logic in the data access code. Thus this also introduces the code
base's first database-backed JUnit test, and to support it, also introduces a
database test runner that brings up a temporary database instance in Docker to
provide consistent test execution everywhere.

There is also another round of data model changes based on spec changes and on
details of the initial implementation.

CU-kgx97u