Use enums for more reference tables
Previously, I had only set the code generator up to make Kotlin enums for reference tables where the business logic would potentially need to vary based on the values. However, using enums also causes the valid values to be documented in the generated OpenAPI docs and shown in the Swagger UI, so convert the others as well.

This also reduces the complexity of database queries that read and write accessions. Future versions of the app might allow users to add values to these lists, in which case we'll have to go back to treating them as dynamic, but for V0 better to keep it simple.