Add endpoint to list storage locations
The front end needs to be able to show the storage condition when the user picks a
storage location from the drop-down menu. And it also needs to know what storage
locations to include in the drop-down menu. Add an endpoint to get the list of
locations and conditions.

CU-m0y9yz