Populate per-site tables based on config file
Add a configuration option `terraware.site-config-url`. At server start time and once an hour after that, the URL will be fetched and its contents applied to the database. The contents should be a JSON-serialized object as described by the `PerSiteConfig` class.

In dev environments, the configuration is read from the classpath by default.

CU-m4yhr0