Add withdrawals field to accession
The GET and PUT endpoints for accessions now include a `withdrawals` field with a list of withdrawals.

Since withdrawals include a server-computed value (seed count), the PUT endpoint is modified to return the updated accession data including server-computed values. This will also be useful for other server-initiated business logic such as state transitions.