Treat MQTT connect failures as non-fatal
Previously, if MQTT was enabled, successfully connecting to the broker was treated
as a required step in the server's start sequence. So if the broker was down, the
server would fail to start.

Instead, retry the broker connection in the background if the initial attempt
fails.

In theory the Paho client library is supposed to do that for us, but in practice
it appears to have a number of bugs in its reconnect logic, so our code now does
its own connection management and tries to clean up after the Paho library's bugs.

CU-m8yaz6[completed]