Add Docker entrypoint to mount disk on Balena
Under Balena, each container needs to mount its own filesystem from the RAID
array. Add an entrypoint script to mount the photos volume to `/data` when running
under Balena, with the usual provisions for overriding the device name and using
a Docker volume instead of an external disk.

This required changing the OpenJDK base image to a Debian one because the default
image didn't include the `mount` command.