Add device information to accession
We want to record information about mobile devices when accessions are created by the seed collector app. Add a field to the accession creation endpoint and return its value when the accession is queried later. The device information can only be set at creation time, not updated, since it is supposed to describe the device that submitted the accession to the server initially.

CU-mjzwxv