Initial version of script to generate test data
It is not very full-featured yet, but add a script to populate a server with test
data so we can start testing search functionality, and a script to do a simple
example search and report how long it takes.

Add some indexes to columns based on the observed performance of these scripts.

CU-mf27bm