Rename "status" to "active" in accession API
The names "state" and "status" were too similar and were easy to confuse with one another. Rename "status" to "active" to differentiate them. The values ("Active" and "Inactive") remain the same. This is a derived value, not explicitly stored in the accession table, so no database migration is needed.

CU-md0gcd[completed]
CU-md0gpa[completed]