Restore AccessionPayload.photoFilenames
I removed this from the request payloads since it doesn't make sense as a writable value and we have a separate API to add photos, but I should not have removed it from the response payload since the front end code uses it to show the list of photos on the accession page. Put it back in `AccessionPayload` and populate it with the actual photo filenames.

CU-mf43gj