Accession state transition logic
Accession states will now update based on client-initiated changes to data, as
well as based on the passage of time.

`AccessionModelTest` includes an explicit test case for every transition rule
in the specification, plus some tests to make sure transitions _don't_ happen
when the spec doesn't say they should. This should document the intended
behavior in addition to catching future bugs.

The `accession_state_history` table is populated when states change. It now
includes a text column to store the reason for the state transition, in
anticipation of people wanting to know why an accession is in the state it's
in.

CU-mt0qmm
CU-mt0qna
CU-mk4bky