Sort enumerated fields by name, not ID
If the user sorts search results by a field whose representation in the database is an ID pointing to a reference table, we want to sort the results by the text value that's shown to the user, not by the ID which is an internal implementation detail.

To support this, modify the generated enums so they implement a common interface the application code can use to examine the underlying representation in a generic way.

CU-mq0au8