Make search use human-readable display name for enum fields
It was returning the enum constant's internal string representation which was the same for single-word values but not for values that contain whitespace. So for example, the "In Storage" state was getting returned as "InStorage".

Since the fix was the same as for the above, also fix rendering of very small or very large decimal values, which used to use scientific notation.

CU-n566mk[review]
CU-n50d0a[review]