Allow configuration of API key
Though we'll eventually want the ability to manage multiple API keys, for now the only consumer of the API is the `site` container in the Balena stack, and it just needs to use an API key that the server will recognize. Add a config option to allow the key to be passed in as an environment variable, creating it in the database as needed.

CU-n2xqa0