Add endpoint to fetch per-site device configuration
Add a new endpoint `/api/v1/device/all/config` to get the list of all the devices at the current site with the configuration information the device manager needs in order to talk to the devices.

This is a superset of the information in the device list CSV file consumed by the device manager code; in addition to the information there, it also includes make, model, and device type fields the device manager can use to determine what metrics to report. Those fields are now set in the per-site configuration JSON file and are mandatory since the device manager will need to use them to drive its device interactions.

Add the device manager endpoints to the Swagger documentation.

CU-nb4hnz