Put seedbank-server git revision in Server header
When the server is pulled from Docker Hub, there's no great way to tell which
revision of the code you're using, which makes bug reports less precise. Generate
a version number based on the current git revision and put it in the `Server`
header line.

The version number is calculated early on in the build process, so it will be
available for other Gradle tasks like publishing artifacts.

CU-n34ahc