Switch to official Kotlin release
Previously we were using a snapshot release to get a fix for a standard library
bug, but (a) it was only affecting Micronaut, and (b) it was included in the
most recent official release.

Also upgrade to a ktfmt version that is built on this Kotlin version (none of the
formatting rule changes in ktfmt affect our code).