Add endpoint to report a timeseries value
To monitor disk usage, we're going to run a little shell script in each Balena container that mounts a filesystem from the SSD. It sends free space statistics to seedbank-server by running curl. Add an endpoint to support that. It accepts an HTML-form-style request instead of JSON to make it more curl-friendly.

CU-p7585k