Update schema terminology to match UI terminology
The names of some concepts evolved as the application specs were fleshed out,
and we were using older names in some cases. Make the names in the database
schema match the user-facing names to reduce the amount of name mapping we have
to do in our heads as we work on the code.

There were also some schema elements that ended up not being needed as the apps
evolved; drop those to reduce clutter.

CU-pk0vdn