Decouple SearchService from API payloads
It's not great practice to tightly couple external API payload formats and the
internal implementation of the business logic. Fix this for the search
endpoints using a combination of explicit method arguments and alternate data
classes that have the semantics the search code needs (which sometimes differ
slightly from the external API, e.g., in regard to nullability).

CU-m0z57b