Decouple geolocation, photo metadata models from controllers
Guiding principle is that the service layer shouldn't take controller-specific
classes as input.

For geolocation, this is just an ownership change: the class moves to the
`model` package.

For photo metadata, define an interface that includes the fields we care about,
and make `PhotoRepository` take that interface as input. Move a couple of bits
of metadata from method arguments to the metadata interface since conceptually
they should be grouped together with the other metadata.

CU-m0z57b