Add source field to accession
Add a computed field to the accession object that the client can look at to
tell whether an accession originally came from the web UI or the seed collector
app.

Right now the logic is very dumb: if there was mobile device information
submitted when the accession was created, it is assumed to have come from the
seed collector app. If/when we start to add more sources of accession data, we
can make that logic more sophisticated.

CU-pv8t51