Ignore read-only accession fields
When the client updates an accession, it may include fields that are returned
as part of the accession data via the accession GET endpoint but aren't
recognized on PUT requests, such as the list of photo filenames. Previously,
this caused the server to reject the PUT request. Now it should ignore them.

The warning about this wasn't being logged by the server, possibly because
there was no default log level, so set that to make future bugs like this easier
to track down.

CU-px8k25