Add "Unsure" option to endangered and rare fields
Update the accession data model to turn `endangered` and `rare` into enumerated
fields rather than booleans.

Since this is a backward-incompatible API change, introduce new accession
endpoints that expose the new enumerated types. The original accession endpoints
continue to accept and return boolean values. This is lossy; writing "Unsure"
to one of those fields using the v2 API and then reading the accession with the
v1 API will cause it to show up as null.

Search is handled similarly: there are two new search fields `endangered2` and
`rare2` that expose the enumerated values, and the existing fields `endangered`
and `rare` return boolean values.

Once the web and mobile clients have been updated to use the v2 API, we will
update the v1 endpoints to also use the enums (getting rid of the boolean
emulation). After that, the clients can switch back to the v1 URLs and we can
get rid of the v2 ones to keep the API namespace tidy.

CU-q32g1g