Upgrade Docker runtime to Java 16
Switch to the Java 16 OpenJDK base image for deployment to the servers. Use the
"Debian slim" version of the image to reduce size; apart from `curl` (which is
now installed explicitly) we weren't using any of the commands or libraries
that are added in the non-slim version.

Note that the code is still built using Java 15, because Gradle does not yet
officially support doing builds using a Java 16 toolchain. Our code is in Kotlin
anyway, so the fact that we can't use Java 16 language features isn't relevant.

CU-p57169