Remove migrations with canned dev environment data
Before we had `PerSiteConfigUpdater`, we used a repeatable database migration
to populate dev environment databases with a few example rows for ease of
testing. But now we have a more robust mechanism to initialize a database, so
the migration's data ends up just being clutter.

A few tests expected the data from the migration; change the test code to
explicitly insert what they need.

CU-qv2k14