Insert API key after updating per-site config
On a completely clean database, it was possible for the server to attempt to
insert the API key before the per-site config JSON file was processed. The
organization for the API key doesn't exist yet in that case, so the key can't
be inserted.

Fix it by updating the key again after the per-site config is written to the
database. But still try to update the key immediately at start time because on
an existing database, we want to be able to rotate the key and have the new key
start working immediately even if the per-site config JSON isn't available.