Use console, not /dev/stdout, for access logs
Update the logging configuration so it sends access logs to the server's
standard output rather than explicitly trying to open `/dev/stdout` under
Balena. That device file isn't accessible in the container if we try to run the
Java process as an unprivileged user.

Get rid of a bunch of Micronaut-specific log configuration left over from
before the switch to Spring Boot.

CU-r3a482