Stop running seedbank-server as root
The seedbank-server container needs to start as root on Balena devices so that
it can mount the photos filesystem from the RAID array. But after that, nothing
requires privileged access. Run the Java process and the metrics script as an
unprivileged user.

Modify the development Docker Compose configuration so it actually runs the
image that gets built locally, so we can test these kinds of changes.

CU-r3a482