Remove transitional rare/endangered fields
Step 3 of a 3-step process for getting rid of some backward-compatibility code
we added when we changed the `rare` and `endangered` accession fields from
booleans to enums. We added new endpoints under `/api/v2` that exposed the
enums, and made the existing `/api/v1` ones simulate the old boolean behavior.

1. Make the v1 endpoints expose the enums once all clients have switched to
   the v2 ones.
2. Update apps to switch back to the v1 endpoints.
3. Remove the v2 endpoints. (That's this change.)

CU-q518wr