Reduce number of sources of truth in build process
There were too many places in the seedbank-server build where information lived
in multiple places and could potentially get out of sync. This is an attempt to
get closer to single-source-of-truth for everything. Specifically:

Code generation (which happens in the `buildSrc` subproject) was using a
slightly different version of the jOOQ library than the application code.
Solve this by moving the jOOQ library version to the project properties rather
than hardwiring it in the build scripts. (For consistency, I moved a few other
version numbers to the properties file at the same time.)

The reference tables that get turned into Kotlin enums were listed in two
places (`EnumGenerator.kt` and `jooq-codegen.xml`). Solve this by switching to
a different jOOQ Gradle plugin that allows the jOOQ configuration to be
constructed in build script code, so we don't need the XML file any more and
all the information about the generated enums is in `EnumGenerator.kt`.

The new plugin puts its generated code in the `build` directory rather than in
`src/generated`, so we no longer need `src/generated` at all.

CU-j4y3u9