Add Nursery state and nursery start date
Add a new accession field `nurseryStartDate` and a new state `Nursery`.

The logic for determining which state an accession is in is no longer a state
machine with a set of possible transitions but is now a value that's derived
purely from the values (or the presence) of various fields on the accession.
This means it is now possible for an accession's state to jump around
arbitrarily depending on which fields users edit.

CU-p91dr5
CU-pv7n72