Implement species management API
Add new `/api/v1/seedbank/values/species` endpoints to allow the client to list,
create, and rename species. Species are still created implicitly as needed during
accession creation (that is, the fact that you _can_ now explicitly create a
species doesn't mean you _have_ to do it).

To support the front end being able to apply a species name edit across all
accessions from the accession editing UI, include the species ID in the accession
payload. That ID can be passed to the API endpoint to edit species data. The ID is
ignored if it's present in an incoming accession payload from the client.

CU-453u9v