Add "simulate" option to accession PUT endpoint
To allow the front end to show the results of a particular accession edit without
actually saving it to the database, add a `simulate` query-string parameter to the
accession PUT endpoint. If it is set to `true`, the server will not save any
changes to the database, but will return the updated accession data that it would
have saved to the database normally. This includes server-generated values such as
the withdrawals that get created for germination tests.

CU-tn5vd2