Support count and weight measurements of seeds
Replace the various seed count and seed weight elements of the data model with a
more generic "quantity" concept that can represent a seed count or a weight
measurement.

An accession's processing method determines which kinds of quantities the system
expects it to have.

The server-side calculations are different for weight-based and count-based
accessions: values such as "quantity remaining in the accession" are calculated by
the server for count-based accessions but are supplied as user input for
weight-based accessions.

The relationship between withdrawals and germination tests is also reworked here:
every germination test whose seed quantity is known is now associated with a
withdrawal. This simplifies the conceptual model; now every removal of seeds from
the accession is a withdrawal, whether it's for testing or for distribution.

CU-td4jrz