Support optional weight units for grams fields
Allow the client to add weight unit suffixes to exact and range search values
for the various "grams" search fields. The server converts the values to grams
internally. This allows the client to specify that, e.g., it wants to match
accessions with between 10 and 20 grams remaining and get back an accession where
the user specified the quantity as 12000 milligrams.

CU-tq9qc2