Separate internal models from API payloads
Initially, as a time-saving shortcut, the API payload classes were passed
directly to the business-logic layer. This is bad separation-of-concerns
practice and it was already starting to cause the code to accumulate ugly
workarounds to avoid name collisions and so forth.

Fix it by explicitly converting between payload classes and internal model
classes, and removing the interfaces that allow the two to be used
interchangeably, and remove a bunch of workarounds that are no longer needed.

CU-v9a7bb