Return more detailed error message if device not found
Currently if the device manager tries to access a timeseries for a nonexistent
device, it's hard to tell why the device doesn't exist: the device name could
be correct but the organization name could be wrong, for example.

Add more detailed diagnostics to make it easier to tell what's wrong.

CU-w5emym