Require users to log in; support user identity
This commit adds user identity and permissions to terraware-server. Previously,
the code was still designed to run on a local server in a seed bank where there
was no requirement to know which users were which.

API requests must now be routed through OAuth2 Proxy, which will require the user
to log into a Keycloak authentication server. The documentation in README.md is
updated to reflect that, and a new file KEYCLOAK.md has instructions for setting
up Keycloak and Oauth2 Proxy in a local dev environment.

See the class docs for `SecurityConfig` for a quick overview of the moving pieces.