Rename project to terraware-server
This code base already covers more than just seed banking, and it is going to
expand further into more areas. Rename it to a terraware-server to make it clear
that it's not a seed-bank-specific back end.

Subsequent commits will remove internal references to the fact that this used to
be called "seedbank-server". For example, the package names still say "seedbank"
at the moment.

CU-13nkzt5