Reorganize package hierarchy based on business functions
Previously, the business logic for the seed bank app was mixed with the logic
for device management and customer information. As the server grows, it will be
useful to be able to ignore code for business functions other than the one
you're working on at the moment. So reorganize the package hierarchy to use a
business-function-first scheme rather than an implementation-layer-first one.

Concretely, that means introducing a `com.terraformation.backend.seedbank`
package and moving all the seed bank code to subpackages under there, and
doing the same for a few other top-level business function types. Generic
infrastructural code is still in implementation-layer-based subpackages under
`com.terraformation.backend`.

Mostly this is just file renames and updates to `import` statements, but in a
few cases there were files that mixed business-function-specific code with
generic code. Those have been split up into multiple files.

CU-13nkzt5