Don't use Kotlin 1.4+ features in code generator code
The code-generation code in `buildSrc` is invoked from within the Gradle
process, which means it uses whatever version of Kotlin is being used
internally by Gradle. Currently that's Kotlin 1.3.x.

The code generator code was using a syntax change in 1.4 (trailing commas in
arguments). That didn't break anything because it's just a syntax change and
doesn't result in 1.3-incompatible class files, but IntelliJ correctly flagged
it as a compatibility problem. Switch to 1.3-compatible syntax so it's less
ambiguous which language version the code is targeting.