Introduce type-safe wrappers for database IDs
Many of our tables have autoincrementing 64-bit integer primary keys. Previously,
these IDs were represented as `Long` in the application code. Using the raw
numeric type runs the risk of accidentally using the wrong kind of ID, e.g.,
passing an accession ID to a function that expects a site ID. This is going to get
more likely as the permissions API comes together since it involves checking
permissions on various kinds of objects.

Reduce the risk by introducing wrapper classes. Now an accession ID is of type
`AccessionId` in the application code, rather than type `Long`. This will prevent
us from ever passing the wrong kind of ID; if you try, it'll fail to compile.

The ID classes are generated by the build, just like the generated enum classes
for reference tables. Since we're now generating more than just enums, refactor
the code generation code in `buildSrc` a bit to make it easier to work with.
The code generator configuration (lists of table and column names) is now in
its own source file rather than mixed in with the generator logic.