Add projects as a layer between sites and orgs
In our early product designs (aka: wild guesses before we understood much about
the requirements) we thought an organization would be associated with a list of
sites. But we now realize that the sites need to be grouped into projects.

Update the code accordingly: the hierarchy now looks like org -> project ->
site -> facility.

There is temporary code in place to allow the server to consume per-site config
files that still use the old hierarchy; it will generate fake projects to glue
the configured sites to the configured organizations.