Parameterize facility ID in internal code
In preparation for supporting a central installation of the seed bank app, stop
treating facility ID as a server-wide config option in the interior parts of the
code base. We still use the config option, but it's now only accessed by the
controller classes so we can replace it with a client-supplied value once we
update the external API to let clients pass it in.

CU-185y6zy