Barebones administration UI
Add a very basic HTML admin UI. It is accessible at `/admin` and requires requests
to be forwarded through OAuth2 Proxy.

This doesn't include any kind of real administration API; the UI is just a series
of form submission handlers. This should be regarded as a stopgap until we're
ready to implement a more robust administration system.