Create and delete API keys from admin UI
The organization view of the placeholder admin UI now includes a list of API keys
and a button to create a new key. Currently, the intended audience for these keys
is people working on the device manager code: the device manager will need to use
these keys to authenticate to terraware-server.

These "API keys" are actually OAuth2 refresh tokens with no expiration date (aka
"offline tokens" in the lingo of OpenID Connect). The device manager will use them
to authenticate to Keycloak and request access tokens. This requires some
additional setup on the Keycloak side, which is documented in detail in
`KEYCLOAK.md`.

Behind the scenes, the code now has the ability to dynamically create and destroy
Keycloak users and to request offline refresh tokens from Keycloak.

CU-185t84y