Remove MQTT support
When terraware-server was running locally on site, it received device metrics via
a local MQTT broker. For reporting metrics to the cloud, we're going to need to
define our own API that plays well with our authentication scheme. So the MQTT
code no longer serves any purpose and can be removed.

Some of the existing device API endpoints still make reference to "MQTT topics"
because the way timeseries are identified piggybacks on the MQTT topic naming
scheme. Most likely we'll move away from that as we design our metrics API, but
for now the existing endpoints continue to work as before.