Add exceptions for Keycloak failures (#14)
Make `UserStore` throw more precise exceptions when Keycloak can't find a user
or a Keycloak request fails.

Add a test suite for `UserStore`. This also adds a stub implementation of
part of the Keycloak admin API; that implementation is used to simulate error
conditions and will get more sophisticated in an upcoming change that adds more
interactions with Keycloak.