Add support for JSON-formatted logfiles (#16)
Currently, terraware-server outputs its logs in human-friendly form even in
staging/production environments. That means consumers of the logs have to parse
the log messages to extract metadata, which can be error-prone in the face of
log messages with embedded newlines such as stack traces.

Add support for writing a logfile in JSON format, with daily rotation. These
logs have all the metadata already split out into JSON objects.

This will also let us attach rich metadata to log messages in the future; with
text-based logs, it can be hard to represent such metadata in a way that's
easy to parse.

JSON logging is activated by adding `jsonlog` to the list of active Spring
profiles, and by setting the `LOG_FILE` environment variable to the full path
of the logfile.