Declare missing DAO beans; add CI test to catch it next time
There were some DAO classes that were declared as dependencies of some managed
beans but weren't ever declared by the code. This caused the server to fail to
start up, though tests continued to pass since none of our tests involves
constructing the entire dependency graph.

Add the missing DAOs, and also add a CI step to detect this kind of problem in
the future: the Gradle task to generate the OpenAPI docs launches a temporary
server instance, and if there's a problem with the bean graph, it will fail.