Implemented Plants API
Implemented create, read, and update endpoints. Included two
batch read endpoints, one of which provides a summary of which
plant species exist in a layer. Did not include a delete endpoint.
If a caller wishes to delete a plant, then they should delete the
entire feature through the Features API.

Overloaded a couple of the layer data permissions checking
functions so that the caller can pass either a layer id or a
feature id.

Changed the Feature and Layer APIs to use plurals - "features" and
"layers". Made the delete feature endpoint throw an error if the
feature does not exist.