Support GeoJSON serialization of Geometry objects
Add support for serializing and deserializing PostGIS geometry classes in GeoJSON
format.

Controllers don't have to do anything special to use this; they can just include a
`Geometry` field in a request or response payload and the serializer will
automatically do the JSON conversion.

A test suite verifies that the JSON is identical to the output of the PostGIS
`ST_AsGeoJSON` function.

With this change, the application code should largely not have to know that
GeoJSON exists at all; it uses `Geometry` objects to talk to both the database and
the client.