FT-126 Support PostGIS GEOMETRY columns (#22)
Add a jOOQ data type binding class so that columns of type `GEOMETRY` can be
accessed from application code in a clean way. Currently, the only column of
that type in the database is `features.geom`.

Previously, `features.geom` was unconstrained: it could hold a mix of 2D and 3D
shapes in a mix of coordinate systems. It was up to the application code to ensure
that the shapes were consistent.

Update the database to require that all shapes be in 3D and in the spherical
Mercator (aka "pseudo-Mercator") coordinate system. Attempts to insert 2D shapes
or shapes in a different coordinate system will be rejected by the database.

As part of that, remove the `altitude` column, copying its value into the Z
coordinate of any existing 2D shapes in the database. A single `altitude` column
meant we were unable to express things like polygons on hillsides where not all 
the points had the same altitude.

It's annoying to have to explicitly convert client input to the right coordinate
system, so make the jOOQ bindings automatically do the transformation when
geometry values are passed as arguments to database queries.

Update `FeatureModel` to include a `geom` field and remove `altitude`.