Add geom property to feature API payloads
The feature API endpoints now accept and return location data in GeoJSON format.

Unfortunately, out of the box, the PostGIS geometry classes cause the OpenAPI
schema generator to bomb out because they overload some property accessor methods
in a way the OpenAPI code doesn't expect. To fix that, explicitly define a schema
and tell the OpenAPI code to use it instead.