Use URLs for file storage locations (#26)
Down the road, we may want to support multiple file storage locations, e.g.,
different S3 buckets for different photo types. In anticipation of that, and
also in anticipation of an upcoming change to store photo file locations in
the `photos` table, change the `FileStore` API so it accepts URLs rather than
paths.

This doesn't actually support multiple active `FileStore`s, but will make it
easier to do so in the future since we'll be able to look at the URL to route
calls to the correct one.