Implemented Plants and Plant Observations API (#24)
--- Main API Changes ---
Implemented create, read, and update endpoints for Plants and
Plant Observations. Plants API includes two somewhat more
involved list endpoints (`/list` and `/list/summary`). Neither API
includes a delete endpoint. If a client wishes to delete a plant 
or a plant observation, then they should delete the entire feature
through the Features API. Especially for the Plant Observations
API, we may want to include delete functionality in the future. 

There are a lot of tests associated with both APIs. Since I'm using
the autogenerated PlantsRow and PlantObservationsRow, there
are more edge cases to handle.

--- Permissions Changes ---
Overloaded a couple of the "layer data" permissions check
functions so that the caller can pass either a layer id or a
feature id.

--- Other API Changes ---
Changed the Feature and Layer APIs to use plurals ("features"
and "layers"). Made the `delete` feature endpoint throw an error
if the feature does not exist. Currently, the permissions check will
also check if the feature exists, but I didn't want to depend on
that in the future.