FT-90 Unify accession and feature photo data
Previously, the `photos` table was specifically for photos of map features, and
there was a separate `accession_photos` table for photos of seeds.

Change the database schema so that `photos` has information about photos of all
varieties. The new table `feature_photos` has feature-specific information about
photos, and `accession_photos` just links photos and accessions.

As part of the unification, `photos` now has a single `location` field for the
GPS data for each photo. It is constrained so that it has to be a PostGIS "Point"
value in spherical Mercator coordinates.

Previously, the storage paths of both accession and feature photos were computed
based on various bits of metadata. But we want a consistent layout for all our
photos. So this change also introduces a `PathGenerator` class that constructs
storage paths based on timestamps rather than on photo-type-specific metadata.
These paths are designed to work with the `FileStore` interface introduced
previously. Paths are stored in the database as part of the photo's metadata
rather than computed on the fly so that any changes to the path generation
algorithm won't cause existing photos to become inaccessible.